+++ DXMainClient/DXGUI/GameClass.cs
@@ -1,5 +1,4 @@
 using ClientCore;
-using ClientCore.CnCNet5;
 using ClientGUI;
 using ClientGUI.IME;
 using DTAClient.Domain;
