+++ DXMainClient/DXGUI/Generic/LoadingScreen.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using ClientCore;
-using ClientCore.CnCNet5;
+using DTAClient.Domain.Multiplayer.CnCNet;
 using ClientCore.Extensions;
 
 using ClientGUI;
