+++ DXMainClient/DXGUI/Multiplayer/CnCNet/CnCNetLoginWindow.cs
@@ -1,5 +1,5 @@
 ﻿using ClientCore;
-using ClientCore.CnCNet5;
+using DTAClient.Domain.Multiplayer.CnCNet;
 using ClientGUI;
 using ClientCore.Extensions;
 using Microsoft.Xna.Framework;
