+++ DXMainClient/DXGUI/Multiplayer/PlayerListBox.cs
@@ -1,4 +1,4 @@
-﻿using ClientCore.CnCNet5;
+﻿using DTAClient.Domain.Multiplayer.CnCNet;
 using DTAClient.Online;
 using ClientCore.Extensions;
 using Microsoft.Xna.Framework;
