+++ DXMainClient/DXMainClient.csproj
@@ -44,6 +44,7 @@
     <PackageReference Include="System.Text.Json" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\ClientGUI\ClientGUI.csproj" />
     <ProjectReference Include="..\DTAConfig\DTAConfig.csproj" />
     <ProjectReference Include="..\ClientUpdater\ClientUpdater.csproj" />
   </ItemGroup>
